requirements.txt
fastapi[all]~=0.111.0
requests
sqlalchemy~=2.0.31
alembic~=1.13.2
asyncpg
pydantic-settings
pydantic~=2.8.0
uvicorn
prometheus-client
psutil
httpx~=0.27.0
python-jose
bcrypt==4.0.1
passlib[bcrypt]
celery~=5.3.6
redis~=4.6.0
flower~=1.2.0
python-dotenv
pillow~=10.3.0
torch~=2.2.0
torchvision~=0.17.0
loguru~=0.7.2
pyopenssl~=24.0.0
cryptography~=42.0.5